support System.clock and seeding for random, fix for importing built in modules
System.clock and Random.new use time based functions from the host. Updated wren to make it easier to set these functions from the host but also remove dependency for "wasi_snapshot_preview1.clock_time_get" and also doesn't have to rely on the deprecated wasi-sdk clock function.

Also fixes an issue where importing wren's built in module would error since the host didn't bind anything to those modules.

+++ internals/f13.go
@@ -1,6 +1,6 @@
 package internals
 
-// module: wasi_snapshot_preview1, field: fd_write
-func f13(ctx *Context, l0 int32, l1 int32, l2 int32, l3 int32) int32 {
+// module: wasi_snapshot_preview1, field: fd_seek
+func f13(ctx *Context, l0 int32, l1 int64, l2 int32, l3 int32) int32 {
 	return ctx.f.F13(ctx, l0, l1, l2, l3)
 }
